backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-pptx==0.6.23
Pillow==10.1.0
opencv-python==4.8.1.78
moviepy==1.0.3
numpy==1.24.3
aiofiles==23.2.1
pydantic==2.5.0










